requirements.txt
aiohttp
aiosignal
altgraph
annotated-types
anyio
attrs
bottle
certifi
cffi
charset-normalizer
click
colorama
docker
FeatureCloud
flake8
frozenlist
gitdb
GitPython
idna
joblib
jsonpickle
macholib
mccabe
multidict
neo4j
numpy
packaging
pandas
propcache
pycodestyle
pycparser
pydot
pyflakes
pyinstaller
pyinstaller-hooks-contrib
pyparsing
python-dateutil
python-dotenv
pytz
PyYAML
requests
scikit-learn
scipy
setuptools
six
smmap
sniffio
threadpoolctl
tqdm
tzdata
urllib3
websocket-client
yarl
jsonnet
pickle5
allennlp
ase
blis
boto3
botocore
catalogue
configparser
cycler
cymem
datasets
dill
docker-pycreds
docutils
et-xmlfile
fastobo
filelock
fonttools
fsspec
gensim
googledrivedownloader
huggingface-hub
isodate
jmespath
kaleido
littleutils
llvmlite
lmdb
matplotlib
more-itertools
multiprocess
murmurhash
nltk
node2vec
numba
ogb
openpyxl
outdated
overrides
pathlib
pathtools
pathy
plotly
preshed
progress
promise
pronto
protobuf
pyarrow
pydantic
pydeprecate
python-igraph
python-louvain
pytorch-lightning
rdflib
regex
s3transfer
sacremoses
seaborn
sentencepiece
sentry-sdk
setproctitle
shortuuid
smart-open
snap-stanford
spacy
spacy-legacy
srsly
subprocess32
tabula-py
tenacity
tensorboardx
termcolor
texttable
thinc
tokenizers
torch
torchmetrics
torchviz
transformers
typer
wandb
wasabi
xlrd
xxhash
yaspin
